% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_loss.R
\name{check_loss}
\alias{check_loss}
\title{Check Loss}
\usage{
check_loss(x, tau)
}
\arguments{
\item{x}{a numeric scalar, vector, or matrix.}

\item{tau}{a numeric scalar between 0 and 1.}
}
\description{
\code{check_loss} computes the check loss function for a given quantile tau.
}
