% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_qr.R
\name{classify_qr}
\alias{classify_qr}
\title{Binary Classifications from Quantile Regression}
\usage{
classify_qr(newx, coef, y_transform, threshold)
}
\arguments{
\item{newx}{a design matrix containing the input values for predictions.}

\item{coef}{a vector or matrix containing the estimated regression coefficient values. 
If a matrix (e.g. if you want classifications for model fits produced by a 
sequence of tuning parameter values), then the each column should correspond
to a different fit.}

\item{y_transform}{a vectorized function to be applied to the predictions.}

\item{threshold}{the threshold to compare the predictions to.}
}
\value{
a vector or matrix of classifications.
}
\description{
\code{classify_qr} creates binary classifications from a quantile
regression fit by comparing predicted values to a prespecified threshold.
}
