% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_woodbury.R
\name{compute_woodbury}
\alias{compute_woodbury}
\title{Woodbury Matrix Inversion Identity}
\usage{
compute_woodbury(mat)
}
\arguments{
\item{mat}{a matrix.}
}
\value{
a matrix, the result of inverting \code{mat}.
}
\description{
\code{compute_woodbury} uses the Woodbury matrix inversion identity to compute
the inverse of (I + M) where M is the cross-product of the design matrix X
and I is the identity matrix.
}
