% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{r_main_parallel_Matrix}
\alias{r_main_parallel_Matrix}
\title{Fit Parallelized QRADMM}
\usage{
r_main_parallel_Matrix(
  X,
  y,
  M,
  intercept,
  max_iter = 500,
  min_iter = 10,
  n_workers,
  lambda,
  tau,
  rho,
  alpha,
  penalty = "lasso",
  penalized,
  inverses_write_path = NULL,
  inverses_read_path = NULL,
  abstol = 1e-07,
  reltol = 1e-04
)
}
\arguments{
\item{M}{an integer specifying the number of data blocks.}

\item{intercept}{a logical indicating whether an intercept should be included in the model.}

\item{max_iter}{an integer, the maximum number of iterations to run.}

\item{min_iter}{an integer, the minimum number of iterations to run.}

\item{n_workers}{an integer, the number of worker processes to create.}

\item{lambda}{a numeric scalar or vector of non-zero values, the values for the tuning parameter.}

\item{tau}{a numeric scalar between 0 and 1, the conditional quantile.}

\item{rho}{a positive numeric scalar.}

\item{alpha}{a positive numeric scalar.}

\item{penalty}{a character string, the penalty to apply.}

\item{penalized}{a logical vector of the same length as the number of model parameters, 
TRUE values indicate that the corresponding parameter should be penalized.}

\item{inverses_write_path}{a character string, the location to save the calculated inverses to.}

\item{inverses_read_path}{a character string, the location to read the calculated inverses from.}

\item{reltol}{}

\item{dat}{a numeric matrix where the first column is the vector of responses and the rest of the columns
are the design matrix for the model.}
}
\description{
\code{r_main_parallel_Matrix} is the model-fitting function for the parallelized QRADMM algorithm
implemented entirely in R.
}
